Id System Revamp & Mersenne Twister RNG
- Added Release Build to tasks in VSCode
- Optimised compilation process

+++ engine/src/math/geometry_3d.h
@@ -15,7 +15,7 @@ typedef enum raycast_hit_type
 typedef struct raycast_hi
  {
     raycast_hit_type type;
-    u32 unique_id;
+    u64 unique_id;
     vec3 position;
     f32 distance;
 } raycast_hit;
