+++ testbed_lib/src/resources/simple_scene.h
@@ -157,6 +157,6 @@ BAPI struct mesh* simple_scene_mesh_get(simple_scene* scene, const char* name);
 BAPI struct skybox* simple_scene_skybox_get(simple_scene* scene, const char* name);
 BAPI struct terrain* simple_scene_terrain_get(simple_scene* scene, const char* name);
 
-BAPI struct transform* simple_scene_transform_get_by_id(simple_scene* scene, u32 unique_id);
+BAPI struct transform* simple_scene_transform_get_by_id(simple_scene* scene, u64 unique_id);
 
 BAPI b8 simple_scene_debug_render_data_query(simple_scene* scene, u32* data_count, struct geometry_render_data** debug_geometries);
